chore: update Shopper.java, ChromeOperator.java, and Searcher.java with no specific changes
